Modify: 파일 이름 변경 및 코드 변경
- vm_product_model.js => vmModel.js
- vm_product.json => vmProduct.json
- vm_product_view.js => vmProductView.js
- vmModel getProducts 메소드 fetch 코드 수정
