+++ vmProductView.js
